+++ spec/rubocop/cop/rails/relative_date_constant_spec.rb
@@ -43,6 +43,23 @@ RSpec.describe RuboCop::Cop::Rails::RelativeDateConstant, :config do
       RUBY
     end
 
+    it 'registers and corrects an offense when a method is chained after a relative date method' do
+      expect_offense(<<~RUBY)
+        class SomeClass
+          START_DATE = 2.weeks.ago.to_date
+          ^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^ Do not assign ago to constants as it will be evaluated only once.
+        end
+      RUBY
+
+      expect_correction(<<~RUBY)
+        class SomeClass
+          def self.start_date
+            2.weeks.ago.to_date
+          end
+        end
+      RUBY
+    end
+
     it 'registers an offense for exclusive end range' do
       expect_offense(<<~RUBY)
         class SomeClass